all: Fixed printing size_t on Windows MinGW in a platform independent way.
Warning: inet/common/ShortBitVector.cc: In constructor 'inet::ShortBitVector::ShortBitVector(unsigned int, unsigned int)':
inet/common/ShortBitVector.cc:47:59: warning: unknown conversion type character 'z' in format [-Wformat=]
   47 |         throw cRuntimeError("size = %u must be less than %zu", size, (size_t)(sizeof(unsigned int) * 8));

See https://stackoverflow.com/questions/1546789/clean-code-to-printf-size-t-in-c-or-nearest-equivalent-of-c99s-z-in-c

+++ src/inet/common/packet/recorder/PcapReader.cc
@@ -36,7 +36,7 @@ void PcapReader::openPcap(const char *filename, const char *packetNameFormat)
     struct pcap_hdr fh;
     auto err = fread(&fh, sizeof(fh), 1, file);
     if (err != 1)
-        throw cRuntimeError("Cannot read fileheader from file '%s', errno is %zu.", filename, err);
+        throw cRuntimeError("Cannot read fileheader from file '%s', errno is %u.", filename, (unsigned int)err);
     if (fh.magic == 0xa1b2c3d4)
         swapByteOrder = false;
     else if (fh.magic == 0xd4c3b2a1)
@@ -62,12 +62,12 @@ std::pair<simtime_t, Packet *> PcapReader::readPacket()
     struct pcaprec_hdr packetHeader;
     auto err = fread(&packetHeader, sizeof(packetHeader), 1, file);
     if (err != 1)
-        throw cRuntimeError("Cannot read packetheader, errno is %zu.", err);
+        throw cRuntimeError("Cannot read packetheader, errno is %u.", (unsigned int)err);
     uint8_t buffer[1 << 16];
     memset(buffer, 0, sizeof(buffer));
     err = fread(buffer, packetHeader.incl_len, 1, file);
     if (err != 1)
-        throw cRuntimeError("Cannot read packet, errno is %zu.", err);
+        throw cRuntimeError("Cannot read packet, errno is %u.", (unsigned int)err);
     if (swapByteOrder) {
         packetHeader.ts_sec = swapByteOrder32(packetHeader.ts_sec);
         packetHeader.ts_usec = swapByteOrder32(packetHeader.ts_usec);
